+++ src/preferences/interaction.rs
@@ -6,4 +6,6 @@ pub struct InteractionPreferences {
     pub twist_duration: f32,
     pub dynamic_twist_speed: bool,
     pub fade_duration: f32,
+
+    pub confirm_discard_only_when_scrambled: bool,
 }